remove : 미사용 모듈, lib 삭제 및 exceptions 디렉토리 이동
