+++ BE/apps/socket-server/src/exceptions/index.ts
