--- BE/apps/socket-server/src/modules/auth/auth.module.ts
@@ -1,6 +1,0 @@
-import { Module } from '@nestjs/common';
-
-@Module({
-  providers: [],
-})
-export class AuthModule {}
