--- BE/libs/interface/src/authenticated.request.interface.ts
@@ -1,9 +1,0 @@
-import { Request } from 'express';
-
-export interface AuthenticatedRequest extends Request {
-  user: {
-    id?: number;
-    email: string;
-    name?: string;
-  };
-}
