--- BE/libs/interface/src/index.ts
@@ -1,1 +1,0 @@
-export * from './authenticated.request.interface';
